agent/pyproject.toml
[project]
name = "coding-loop-agent"
version = "0.1.0"
description = "Autonomous personal agent container: talks to claude-bridge on the host."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    # Required by tests/test_bridge_client.py which imports the real bridge app
    # from ../claude-bridge/src via sys.path manipulation.
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
